twc_experiment/requirements.txt
numpy
pandas
scipy
ipython
matplotlib
nltk
pillow
gym==0.15.4
textworld==1.2.0
tqdm
torch
torchvision
absl-py
python-igraph
transformers
editdistance
